implement test directory structure
