+++ test/MinterSuiteMinters/MinterDAExp/MinterDAExpV1.PRTNR.test.ts
@@ -12,32 +12,30 @@ import { ethers } from "hardhat";
 import EthersAdapter from "@gnosis.pm/safe-ethers-lib";
 import Safe from "@gnosis.pm/safe-core-sdk";
 import { SafeTransactionDataPartial } from "@gnosis.pm/safe-core-sdk-types";
-import { getGnosisSafe } from "./util/GnosisSafeNetwork";
+import { getGnosisSafe } from "../../util/GnosisSafeNetwork";
 
 /**
  * These tests intended to ensure this Filtered Minter integrates properly with
- * V1 core contract.
+ * V2 PRTNR core contract.
  */
-describe("MinterDALinV1_V1Core", async function () {
+describe("MinterDAExpV1_V2PRTNRCore", async function () {
   const name = "Non Fungible Token";
   const symbol = "NFT";
 
-  const firstTokenId = new BN("30000000");
-  const secondTokenId = new BN("3000001");
-
-  const startingPrice = ethers.utils.parseEther("1");
+  const startingPrice = ethers.utils.parseEther("10");
   const higherPricePerTokenInWei = startingPrice.add(
     ethers.utils.parseEther("0.1")
   );
   // purposefully different price per token on core contract (tracked separately)
   const basePrice = ethers.utils.parseEther("0.05");
 
-  const projectOne = 3; // V1 core starts at project 3
+  const projectOne = 0;
 
   const ONE_MINUTE = 60;
   const ONE_HOUR = ONE_MINUTE * 60;
   const ONE_DAY = ONE_HOUR * 24;
 
+  const defaultHalfLife = ONE_HOUR / 2;
   const auctionStartTimeOffset = ONE_HOUR;
 
   beforeEach(async function () {
@@ -56,8 +54,10 @@ describe("MinterDALinV1_V1Core", async function () {
     );
     this.randomizer = await randomizerFactory.deploy();
 
-    const artblocksFactory = await ethers.getContractFactory("GenArt721CoreV1");
-    this.token = await artblocksFactory
+    const coreFactory = await ethers.getContractFactory(
+      "GenArt721CoreV2_PRTNR"
+    );
+    this.token = await coreFactory
       .connect(deployer)
       .deploy(name, symbol, this.randomizer.address);
 
@@ -66,7 +66,7 @@ describe("MinterDALinV1_V1Core", async function () {
     );
     this.minterFilter = await minterFilterFactory.deploy(this.token.address);
 
-    const minterFactory = await ethers.getContractFactory("MinterDALinV1");
+    const minterFactory = await ethers.getContractFactory("MinterDAExpV1");
     this.minter = await minterFactory.deploy(
       this.token.address,
       this.minterFilter.address
@@ -74,7 +74,7 @@ describe("MinterDALinV1_V1Core", async function () {
 
     await this.token
       .connect(deployer)
-      .addProject("project1", this.accounts.artist.address, 0, false);
+      .addProject("project1", this.accounts.artist.address, 0);
 
     await this.token.connect(deployer).toggleProjectIsActive(projectOne);
 
@@ -113,7 +113,7 @@ describe("MinterDALinV1_V1Core", async function () {
       .setAuctionDetails(
         projectOne,
         this.startTime + auctionStartTimeOffset,
-        this.startTime + auctionStartTimeOffset + ONE_HOUR * 2,
+        defaultHalfLife,
         startingPrice,
         basePrice
       );
@@ -123,7 +123,7 @@ describe("MinterDALinV1_V1Core", async function () {
   describe("constructor", async function () {
     it("reverts when given incorrect minter filter and core addresses", async function () {
       const artblocksFactory = await ethers.getContractFactory(
-        "GenArt721CoreV1"
+        "GenArt721CoreV3"
       );
       const token2 = await artblocksFactory
         .connect(this.accounts.deployer)
@@ -158,24 +158,15 @@ describe("MinterDALinV1_V1Core", async function () {
     });
 
     it("calculates the price correctly", async function () {
-      await ethers.provider.send("evm_mine", [
-        this.startTime + auctionStartTimeOffset,
-      ]);
-
-      const step = ONE_MINUTE * 8; // 480 seconds
-      const numSteps = 15;
-      for (let i = 1; i < numSteps; i++) {
+      for (let i = 1; i <= 5; i++) {
         let ownerBalance = await this.accounts.owner.getBalance();
-        let a = ethers.BigNumber.from(i * step).mul(
-          startingPrice.sub(basePrice).toString()
-        );
-        let t = ethers.BigNumber.from(a.toString());
-        let price = startingPrice.sub(t.div(step * numSteps));
-        let contractPriceInfo = await this.minter
-          .connect(this.accounts.owner)
-          .getPriceInfo(projectOne);
-        await ethers.provider.send("evm_mine", [
-          this.startTime + auctionStartTimeOffset + i * step,
+        let price = startingPrice;
+        for (let j = 0; j < i; j++) {
+          price = price.div(2);
+        }
+
+        await ethers.provider.send("evm_setNextBlockTimestamp", [
+          this.startTime + auctionStartTimeOffset + i * defaultHalfLife,
         ]);
         await this.minter.connect(this.accounts.owner).purchase(projectOne, {
           value: price.toString(),
@@ -193,8 +184,7 @@ describe("MinterDALinV1_V1Core", async function () {
         let ownerDelta = (await this.accounts.owner.getBalance()).sub(
           ownerBalance
         );
-        expect(ownerDelta.mul("-1").lte(contractPriceInfo.tokenPriceInWei)).to
-          .be.true;
+        expect(ownerDelta.mul("-1").lte(price)).to.be.true;
       }
     });
 
@@ -206,8 +196,8 @@ describe("MinterDALinV1_V1Core", async function () {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           projectOne,
-          this.startTime + ONE_HOUR,
-          this.startTime + 2 * ONE_HOUR,
+          this.startTime + auctionStartTimeOffset,
+          defaultHalfLife,
           startingPrice,
           basePrice
         );
@@ -226,15 +216,13 @@ describe("MinterDALinV1_V1Core", async function () {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           projectOne,
-          this.startTime + ONE_HOUR,
-          this.startTime + 2 * ONE_HOUR,
+          this.startTime + auctionStartTimeOffset,
+          defaultHalfLife,
           startingPrice,
           basePrice
         );
 
-      await ethers.provider.send("evm_mine", [
-        this.startTime + auctionStartTimeOffset + 2 * ONE_HOUR,
-      ]);
+      await ethers.provider.send("evm_mine", [this.startTime + 5 * ONE_HOUR]);
 
       let contractPriceInfo = await this.minter
         .connect(this.accounts.owner)
@@ -257,14 +245,13 @@ describe("MinterDALinV1_V1Core", async function () {
 
       const receipt = await ethers.provider.getTransactionReceipt(tx.hash);
       const txCost = receipt.effectiveGasPrice.mul(receipt.gasUsed).toString();
-
       console.log(
-        "Gas cost for a successful Linear DA mint: ",
+        "Gas cost for a successful Exponential DA mint: ",
         ethers.utils.formatUnits(txCost, "ether").toString(),
         "ETH"
       );
 
-      expect(txCost.toString()).to.equal(ethers.utils.parseEther("0.0373481")); // assuming a cost of 100 GWEI
+      expect(txCost.toString()).to.equal(ethers.utils.parseEther("0.0228514")); // assuming a cost of 100 GWEI
     });
   });
 
@@ -310,14 +297,16 @@ describe("MinterDALinV1_V1Core", async function () {
 
   describe("setAuctionDetails", async function () {
     it("cannot be modified mid-auction", async function () {
-      await ethers.provider.send("evm_mine", [this.startTime + ONE_HOUR]);
+      await ethers.provider.send("evm_mine", [
+        this.startTime + 2 * auctionStartTimeOffset,
+      ]);
       await expectRevert(
         this.minter
           .connect(this.accounts.artist)
           .setAuctionDetails(
             projectOne,
-            this.startTime + ONE_MINUTE,
-            this.startTime + 2 * ONE_HOUR,
+            this.startTime + auctionStartTimeOffset,
+            defaultHalfLife,
             startingPrice,
             basePrice
           ),
@@ -333,8 +322,8 @@ describe("MinterDALinV1_V1Core", async function () {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           projectOne,
-          this.startTime + ONE_MINUTE,
-          this.startTime + 2 * ONE_HOUR,
+          this.startTime + auctionStartTimeOffset,
+          defaultHalfLife,
           startingPrice,
           basePrice
         );
@@ -349,8 +338,8 @@ describe("MinterDALinV1_V1Core", async function () {
           .connect(this.accounts.additional)
           .setAuctionDetails(
             projectOne,
-            this.startTime + ONE_MINUTE,
-            this.startTime + 2 * ONE_HOUR,
+            this.startTime + auctionStartTimeOffset,
+            defaultHalfLife,
             startingPrice,
             basePrice
           ),
@@ -365,8 +354,8 @@ describe("MinterDALinV1_V1Core", async function () {
           .connect(this.accounts.deployer)
           .setAuctionDetails(
             projectOne,
-            this.startTime + ONE_MINUTE,
-            this.startTime + 2 * ONE_HOUR,
+            this.startTime + auctionStartTimeOffset,
+            defaultHalfLife,
             startingPrice,
             basePrice
           ),
@@ -383,8 +372,8 @@ describe("MinterDALinV1_V1Core", async function () {
           .connect(this.accounts.artist)
           .setAuctionDetails(
             projectOne,
-            this.startTime + ONE_MINUTE,
-            this.startTime + 2 * ONE_HOUR,
+            this.startTime + auctionStartTimeOffset,
+            defaultHalfLife,
             basePrice,
             startingPrice
           ),
@@ -450,63 +439,78 @@ describe("MinterDALinV1_V1Core", async function () {
     });
   });
 
-  describe("enforce and broadcasts min auction length", async function () {
-    it("enforces min/max auction length constraint", async function () {
-      await this.minter
-        .connect(this.accounts.deployer)
-        .resetAuctionDetails(projectOne);
-      // expect revert when creating a new project with min/max reversed
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.artist)
-          .setAuctionDetails(
-            projectOne,
-            this.startTime + ONE_HOUR * 2,
-            this.startTime + ONE_HOUR,
-            startingPrice,
-            basePrice
-          ),
-        "Auction end must be greater than auction start"
-      );
-    });
-
-    it("enforces min auction length constraint", async function () {
+  describe("enforce and broadcasts auction half-life", async function () {
+    it("enforces half-life min/max constraint", async function () {
       await this.minter
         .connect(this.accounts.deployer)
         .resetAuctionDetails(projectOne);
       // expect revert when creating a new project with
-      const invalidLengthSeconds = 60;
+      const invalidHalfLifeSecondsMin = ONE_MINUTE;
       await expectRevert(
         this.minter
           .connect(this.accounts.artist)
           .setAuctionDetails(
             projectOne,
-            this.startTime + ONE_HOUR,
-            this.startTime + ONE_HOUR + invalidLengthSeconds,
+            this.startTime + auctionStartTimeOffset,
+            invalidHalfLifeSecondsMin,
             startingPrice,
             basePrice
           ),
-        "Auction length must be at least minimumAuctionLengthSeconds"
+        "Price decay half life must fall between min and max allowable values"
+      );
+
+      // expect revert when creating a new project with
+      const invalidHalfLifeSecondsMax = ONE_DAY;
+      await expectRevert(
+        this.minter
+          .connect(this.accounts.artist)
+          .setAuctionDetails(
+            projectOne,
+            this.startTime + auctionStartTimeOffset,
+            invalidHalfLifeSecondsMax,
+            startingPrice,
+            basePrice
+          ),
+        "Price decay half life must fall between min and max allowable values"
       );
     });
 
-    it("emits event when min auction length is updated", async function () {
-      const newLengthSeconds = 3601;
-      // emits event when minimum auction length is updated
+    it("emits event when allowable half life range is updated", async function () {
+      const newMinSeconds = 60;
+      const newMaxSeconds = 6000;
+      // emits event when allowable half life range is updated
       await expect(
         this.minter
           .connect(this.accounts.deployer)
-          .setMinimumAuctionLengthSeconds(newLengthSeconds)
+          .setAllowablePriceDecayHalfLifeRangeSeconds(
+            newMinSeconds,
+            newMaxSeconds
+          )
       )
-        .to.emit(this.minter, "MinimumAuctionLengthSecondsUpdated")
-        .withArgs(newLengthSeconds);
+        .to.emit(this.minter, "AuctionHalfLifeRangeSecondsUpdated")
+        .withArgs(newMinSeconds, newMaxSeconds);
     });
 
-    it("validate setMinimumAuctionLengthSeconds ACL", async function () {
+    it("validate setAllowablePriceDecayHalfLifeRangeSeconds guards", async function () {
+      await expectRevert(
+        this.minter
+          .connect(this.accounts.deployer)
+          .setAllowablePriceDecayHalfLifeRangeSeconds(600, 60),
+        "Maximum half life must be greater than minimum"
+      );
+      await expectRevert(
+        this.minter
+          .connect(this.accounts.deployer)
+          .setAllowablePriceDecayHalfLifeRangeSeconds(0, 600),
+        "Half life of zero not allowed"
+      );
+    });
+
+    it("validate setAllowablePriceDecayHalfLifeRangeSeconds ACL", async function () {
       await expectRevert(
         this.minter
           .connect(this.accounts.additional)
-          .setMinimumAuctionLengthSeconds(600),
+          .setAllowablePriceDecayHalfLifeRangeSeconds(60, 600),
         "Only Core whitelisted"
       );
     });
@@ -552,7 +556,6 @@ describe("MinterDALinV1_V1Core", async function () {
 
   describe("reentrancy attack", async function () {
     it("does not allow reentrant purchaseTo", async function () {
-      // advance to time when auction is active
       await ethers.provider.send("evm_mine", [
         this.startTime + auctionStartTimeOffset,
       ]);
