+++ test/MinterSuiteMinters/MinterDALin/MinterDALinV0.test.ts
@@ -12,30 +12,32 @@ import { ethers } from "hardhat";
 import EthersAdapter from "@gnosis.pm/safe-ethers-lib";
 import Safe from "@gnosis.pm/safe-core-sdk";
 import { SafeTransactionDataPartial } from "@gnosis.pm/safe-core-sdk-types";
-import { getGnosisSafe } from "./util/GnosisSafeNetwork";
+import { getGnosisSafe } from "../../util/GnosisSafeNetwork";
 
 /**
  * These tests intended to ensure this Filtered Minter integrates properly with
- * V2 PRTNR core contract.
+ * V1 core contract.
  */
-describe("MinterDAExpV1_V2PRTNRCore", async function () {
+describe("MinterDALinV0_V1Core", async function () {
   const name = "Non Fungible Token";
   const symbol = "NFT";
 
-  const startingPrice = ethers.utils.parseEther("10");
+  const firstTokenId = new BN("30000000");
+  const secondTokenId = new BN("3000001");
+
+  const startingPrice = ethers.utils.parseEther("1");
   const higherPricePerTokenInWei = startingPrice.add(
     ethers.utils.parseEther("0.1")
   );
   // purposefully different price per token on core contract (tracked separately)
   const basePrice = ethers.utils.parseEther("0.05");
 
-  const projectOne = 0;
+  const projectOne = 3; // V1 core starts at project 3
 
   const ONE_MINUTE = 60;
   const ONE_HOUR = ONE_MINUTE * 60;
   const ONE_DAY = ONE_HOUR * 24;
 
-  const defaultHalfLife = ONE_HOUR / 2;
   const auctionStartTimeOffset = ONE_HOUR;
 
   beforeEach(async function () {
@@ -54,10 +56,8 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
     );
     this.randomizer = await randomizerFactory.deploy();
 
-    const coreFactory = await ethers.getContractFactory(
-      "GenArt721CoreV2_PRTNR"
-    );
-    this.token = await coreFactory
+    const artblocksFactory = await ethers.getContractFactory("GenArt721CoreV1");
+    this.token = await artblocksFactory
       .connect(deployer)
       .deploy(name, symbol, this.randomizer.address);
 
@@ -66,7 +66,7 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
     );
     this.minterFilter = await minterFilterFactory.deploy(this.token.address);
 
-    const minterFactory = await ethers.getContractFactory("MinterDAExpV1");
+    const minterFactory = await ethers.getContractFactory("MinterDALinV0");
     this.minter = await minterFactory.deploy(
       this.token.address,
       this.minterFilter.address
@@ -74,7 +74,7 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
 
     await this.token
       .connect(deployer)
-      .addProject("project1", this.accounts.artist.address, 0);
+      .addProject("project1", this.accounts.artist.address, 0, false);
 
     await this.token.connect(deployer).toggleProjectIsActive(projectOne);
 
@@ -113,7 +113,7 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
       .setAuctionDetails(
         projectOne,
         this.startTime + auctionStartTimeOffset,
-        defaultHalfLife,
+        this.startTime + auctionStartTimeOffset + ONE_HOUR * 2,
         startingPrice,
         basePrice
       );
@@ -123,7 +123,7 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
   describe("constructor", async function () {
     it("reverts when given incorrect minter filter and core addresses", async function () {
       const artblocksFactory = await ethers.getContractFactory(
-        "GenArt721CoreV3"
+        "GenArt721CoreV1"
       );
       const token2 = await artblocksFactory
         .connect(this.accounts.deployer)
@@ -135,7 +135,7 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
       const minterFilter = await minterFilterFactory.deploy(token2.address);
 
       const minterFactory = await ethers.getContractFactory(
-        "MinterSetPriceERC20V1"
+        "MinterSetPriceERC20V0"
       );
       // fails when combine new minterFilter with the old token in constructor
       await expectRevert(
@@ -158,15 +158,24 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
     });
 
     it("calculates the price correctly", async function () {
-      for (let i = 1; i <= 5; i++) {
-        let ownerBalance = await this.accounts.owner.getBalance();
-        let price = startingPrice;
-        for (let j = 0; j < i; j++) {
-          price = price.div(2);
-        }
+      await ethers.provider.send("evm_mine", [
+        this.startTime + auctionStartTimeOffset,
+      ]);
 
-        await ethers.provider.send("evm_setNextBlockTimestamp", [
-          this.startTime + auctionStartTimeOffset + i * defaultHalfLife,
+      const step = ONE_MINUTE * 8; // 480 seconds
+      const numSteps = 15;
+      for (let i = 1; i < numSteps; i++) {
+        let ownerBalance = await this.accounts.owner.getBalance();
+        let a = ethers.BigNumber.from(i * step).mul(
+          startingPrice.sub(basePrice).toString()
+        );
+        let t = ethers.BigNumber.from(a.toString());
+        let price = startingPrice.sub(t.div(step * numSteps));
+        let contractPriceInfo = await this.minter
+          .connect(this.accounts.owner)
+          .getPriceInfo(projectOne);
+        await ethers.provider.send("evm_mine", [
+          this.startTime + auctionStartTimeOffset + i * step,
         ]);
         await this.minter.connect(this.accounts.owner).purchase(projectOne, {
           value: price.toString(),
@@ -184,7 +193,8 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
         let ownerDelta = (await this.accounts.owner.getBalance()).sub(
           ownerBalance
         );
-        expect(ownerDelta.mul("-1").lte(price)).to.be.true;
+        expect(ownerDelta.mul("-1").lte(contractPriceInfo.tokenPriceInWei)).to
+          .be.true;
       }
     });
 
@@ -196,8 +206,8 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           projectOne,
-          this.startTime + auctionStartTimeOffset,
-          defaultHalfLife,
+          this.startTime + ONE_HOUR,
+          this.startTime + 2 * ONE_HOUR,
           startingPrice,
           basePrice
         );
@@ -216,13 +226,15 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           projectOne,
-          this.startTime + auctionStartTimeOffset,
-          defaultHalfLife,
+          this.startTime + ONE_HOUR,
+          this.startTime + 2 * ONE_HOUR,
           startingPrice,
           basePrice
         );
 
-      await ethers.provider.send("evm_mine", [this.startTime + 5 * ONE_HOUR]);
+      await ethers.provider.send("evm_mine", [
+        this.startTime + auctionStartTimeOffset + 2 * ONE_HOUR,
+      ]);
 
       let contractPriceInfo = await this.minter
         .connect(this.accounts.owner)
@@ -245,13 +257,14 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
 
       const receipt = await ethers.provider.getTransactionReceipt(tx.hash);
       const txCost = receipt.effectiveGasPrice.mul(receipt.gasUsed).toString();
+
       console.log(
-        "Gas cost for a successful Exponential DA mint: ",
+        "Gas cost for a successful Linear DA mint: ",
         ethers.utils.formatUnits(txCost, "ether").toString(),
         "ETH"
       );
 
-      expect(txCost.toString()).to.equal(ethers.utils.parseEther("0.0228514")); // assuming a cost of 100 GWEI
+      expect(txCost.toString()).to.equal(ethers.utils.parseEther("0.0380046")); // assuming a cost of 100 GWEI
     });
   });
 
@@ -267,46 +280,59 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
         });
     });
 
-    it("does not support toggling of `purchaseToDisabled`", async function () {
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.artist)
-          .togglePurchaseToDisabled(projectOne),
-        "Action not supported"
-      );
-      // still allows `purchaseTo`.
+    it("disallows `purchaseTo` if disallowed explicitly", async function () {
       await ethers.provider.send("evm_mine", [
         this.startTime + auctionStartTimeOffset,
       ]);
       await this.minter
+        .connect(this.accounts.deployer)
+        .togglePurchaseToDisabled(projectOne);
+      await expectRevert(
+        this.minter
+          .connect(this.accounts.owner)
+          .purchaseTo(this.accounts.additional.address, projectOne, {
+            value: startingPrice,
+          }),
+        "No `purchaseTo` Allowed"
+      );
+      // still allows `purchaseTo` if destination matches sender.
+      await this.minter
         .connect(this.accounts.owner)
-        .purchaseTo(this.accounts.artist.address, projectOne, {
+        .purchaseTo(this.accounts.owner.address, projectOne, {
           value: startingPrice,
         });
     });
 
-    it("doesn't support `purchaseTo` toggling", async function () {
-      await expectRevert(
+    it("emits event when `purchaseTo` is toggled", async function () {
+      // emits true when changed from initial value of false
+      await expect(
         this.minter
-          .connect(this.accounts.artist)
-          .togglePurchaseToDisabled(projectOne),
-        "Action not supported"
-      );
+          .connect(this.accounts.deployer)
+          .togglePurchaseToDisabled(projectOne)
+      )
+        .to.emit(this.minter, "PurchaseToDisabledUpdated")
+        .withArgs(projectOne, true);
+      // emits false when changed from initial value of true
+      await expect(
+        this.minter
+          .connect(this.accounts.deployer)
+          .togglePurchaseToDisabled(projectOne)
+      )
+        .to.emit(this.minter, "PurchaseToDisabledUpdated")
+        .withArgs(projectOne, false);
     });
   });
 
   describe("setAuctionDetails", async function () {
     it("cannot be modified mid-auction", async function () {
-      await ethers.provider.send("evm_mine", [
-        this.startTime + 2 * auctionStartTimeOffset,
-      ]);
+      await ethers.provider.send("evm_mine", [this.startTime + ONE_HOUR]);
       await expectRevert(
         this.minter
           .connect(this.accounts.artist)
           .setAuctionDetails(
             projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
+            this.startTime + ONE_MINUTE,
+            this.startTime + 2 * ONE_HOUR,
             startingPrice,
             basePrice
           ),
@@ -322,8 +348,8 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           projectOne,
-          this.startTime + auctionStartTimeOffset,
-          defaultHalfLife,
+          this.startTime + ONE_MINUTE,
+          this.startTime + 2 * ONE_HOUR,
           startingPrice,
           basePrice
         );
@@ -338,8 +364,8 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
           .connect(this.accounts.additional)
           .setAuctionDetails(
             projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
+            this.startTime + ONE_MINUTE,
+            this.startTime + 2 * ONE_HOUR,
             startingPrice,
             basePrice
           ),
@@ -354,8 +380,8 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
           .connect(this.accounts.deployer)
           .setAuctionDetails(
             projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
+            this.startTime + ONE_MINUTE,
+            this.startTime + 2 * ONE_HOUR,
             startingPrice,
             basePrice
           ),
@@ -372,8 +398,8 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
           .connect(this.accounts.artist)
           .setAuctionDetails(
             projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
+            this.startTime + ONE_MINUTE,
+            this.startTime + 2 * ONE_HOUR,
             basePrice,
             startingPrice
           ),
@@ -439,78 +465,63 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
     });
   });
 
-  describe("enforce and broadcasts auction half-life", async function () {
-    it("enforces half-life min/max constraint", async function () {
+  describe("enforce and broadcasts min auction length", async function () {
+    it("enforces min/max auction length constraint", async function () {
+      await this.minter
+        .connect(this.accounts.deployer)
+        .resetAuctionDetails(projectOne);
+      // expect revert when creating a new project with min/max reversed
+      await expectRevert(
+        this.minter
+          .connect(this.accounts.artist)
+          .setAuctionDetails(
+            projectOne,
+            this.startTime + ONE_HOUR * 2,
+            this.startTime + ONE_HOUR,
+            startingPrice,
+            basePrice
+          ),
+        "Auction end must be greater than auction start"
+      );
+    });
+
+    it("enforces min auction length constraint", async function () {
       await this.minter
         .connect(this.accounts.deployer)
         .resetAuctionDetails(projectOne);
       // expect revert when creating a new project with
-      const invalidHalfLifeSecondsMin = ONE_MINUTE;
+      const invalidLengthSeconds = 60;
       await expectRevert(
         this.minter
           .connect(this.accounts.artist)
           .setAuctionDetails(
             projectOne,
-            this.startTime + auctionStartTimeOffset,
-            invalidHalfLifeSecondsMin,
+            this.startTime + ONE_HOUR,
+            this.startTime + ONE_HOUR + invalidLengthSeconds,
             startingPrice,
             basePrice
           ),
-        "Price decay half life must fall between min and max allowable values"
-      );
-
-      // expect revert when creating a new project with
-      const invalidHalfLifeSecondsMax = ONE_DAY;
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.artist)
-          .setAuctionDetails(
-            projectOne,
-            this.startTime + auctionStartTimeOffset,
-            invalidHalfLifeSecondsMax,
-            startingPrice,
-            basePrice
-          ),
-        "Price decay half life must fall between min and max allowable values"
+        "Auction length must be at least minimumAuctionLengthSeconds"
       );
     });
 
-    it("emits event when allowable half life range is updated", async function () {
-      const newMinSeconds = 60;
-      const newMaxSeconds = 6000;
-      // emits event when allowable half life range is updated
+    it("emits event when min auction length is updated", async function () {
+      const newLengthSeconds = 3601;
+      // emits event when minimum auction length is updated
       await expect(
         this.minter
           .connect(this.accounts.deployer)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(
-            newMinSeconds,
-            newMaxSeconds
-          )
+          .setMinimumAuctionLengthSeconds(newLengthSeconds)
       )
-        .to.emit(this.minter, "AuctionHalfLifeRangeSecondsUpdated")
-        .withArgs(newMinSeconds, newMaxSeconds);
+        .to.emit(this.minter, "MinimumAuctionLengthSecondsUpdated")
+        .withArgs(newLengthSeconds);
     });
 
-    it("validate setAllowablePriceDecayHalfLifeRangeSeconds guards", async function () {
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.deployer)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(600, 60),
-        "Maximum half life must be greater than minimum"
-      );
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.deployer)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(0, 600),
-        "Half life of zero not allowed"
-      );
-    });
-
-    it("validate setAllowablePriceDecayHalfLifeRangeSeconds ACL", async function () {
+    it("validate setMinimumAuctionLengthSeconds ACL", async function () {
       await expectRevert(
         this.minter
           .connect(this.accounts.additional)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(60, 600),
+          .setMinimumAuctionLengthSeconds(600),
         "Only Core whitelisted"
       );
     });
@@ -556,6 +567,11 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
 
   describe("reentrancy attack", async function () {
     it("does not allow reentrant purchaseTo", async function () {
+      // admin allows contract buys
+      await this.minter
+        .connect(this.accounts.deployer)
+        .toggleContractMintable(projectOne);
+      // advance to time when auction is active
       await ethers.provider.send("evm_mine", [
         this.startTime + auctionStartTimeOffset,
       ]);
@@ -606,6 +622,10 @@ describe("MinterDAExpV1_V2PRTNRCore", async function () {
 
   describe("gnosis safe", async function () {
     it("allows gnosis safe to purchase in ETH", async function () {
+      // admin allows contract buys
+      await this.minter
+        .connect(this.accounts.deployer)
+        .toggleContractMintable(projectOne);
       // advance to time when auction is active
       await ethers.provider.send("evm_mine", [
         this.startTime + auctionStartTimeOffset,
