+++ test/MinterSuiteMinters/MinterDALin/MinterDALinV1.test.ts
@@ -12,20 +12,20 @@ import { ethers } from "hardhat";
 import EthersAdapter from "@gnosis.pm/safe-ethers-lib";
 import Safe from "@gnosis.pm/safe-core-sdk";
 import { SafeTransactionDataPartial } from "@gnosis.pm/safe-core-sdk-types";
-import { getGnosisSafe } from "./util/GnosisSafeNetwork";
+import { getGnosisSafe } from "../../util/GnosisSafeNetwork";
 
 /**
  * These tests intended to ensure this Filtered Minter integrates properly with
  * V1 core contract.
  */
-describe("MinterDAExpV1_V1Core", async function () {
+describe("MinterDALinV1_V1Core", async function () {
   const name = "Non Fungible Token";
   const symbol = "NFT";
 
   const firstTokenId = new BN("30000000");
   const secondTokenId = new BN("3000001");
 
-  const startingPrice = ethers.utils.parseEther("10");
+  const startingPrice = ethers.utils.parseEther("1");
   const higherPricePerTokenInWei = startingPrice.add(
     ethers.utils.parseEther("0.1")
   );
@@ -38,7 +38,6 @@ describe("MinterDAExpV1_V1Core", async function () {
   const ONE_HOUR = ONE_MINUTE * 60;
   const ONE_DAY = ONE_HOUR * 24;
 
-  const defaultHalfLife = ONE_HOUR / 2;
   const auctionStartTimeOffset = ONE_HOUR;
 
   beforeEach(async function () {
@@ -67,7 +66,7 @@ describe("MinterDAExpV1_V1Core", async function () {
     );
     this.minterFilter = await minterFilterFactory.deploy(this.token.address);
 
-    const minterFactory = await ethers.getContractFactory("MinterDAExpV1");
+    const minterFactory = await ethers.getContractFactory("MinterDALinV1");
     this.minter = await minterFactory.deploy(
       this.token.address,
       this.minterFilter.address
@@ -114,7 +113,7 @@ describe("MinterDAExpV1_V1Core", async function () {
       .setAuctionDetails(
         projectOne,
         this.startTime + auctionStartTimeOffset,
-        defaultHalfLife,
+        this.startTime + auctionStartTimeOffset + ONE_HOUR * 2,
         startingPrice,
         basePrice
       );
@@ -124,7 +123,7 @@ describe("MinterDAExpV1_V1Core", async function () {
   describe("constructor", async function () {
     it("reverts when given incorrect minter filter and core addresses", async function () {
       const artblocksFactory = await ethers.getContractFactory(
-        "GenArt721CoreV3"
+        "GenArt721CoreV1"
       );
       const token2 = await artblocksFactory
         .connect(this.accounts.deployer)
@@ -159,15 +158,24 @@ describe("MinterDAExpV1_V1Core", async function () {
     });
 
     it("calculates the price correctly", async function () {
-      for (let i = 1; i <= 5; i++) {
-        let ownerBalance = await this.accounts.owner.getBalance();
-        let price = startingPrice;
-        for (let j = 0; j < i; j++) {
-          price = price.div(2);
-        }
+      await ethers.provider.send("evm_mine", [
+        this.startTime + auctionStartTimeOffset,
+      ]);
 
-        await ethers.provider.send("evm_setNextBlockTimestamp", [
-          this.startTime + auctionStartTimeOffset + i * defaultHalfLife,
+      const step = ONE_MINUTE * 8; // 480 seconds
+      const numSteps = 15;
+      for (let i = 1; i < numSteps; i++) {
+        let ownerBalance = await this.accounts.owner.getBalance();
+        let a = ethers.BigNumber.from(i * step).mul(
+          startingPrice.sub(basePrice).toString()
+        );
+        let t = ethers.BigNumber.from(a.toString());
+        let price = startingPrice.sub(t.div(step * numSteps));
+        let contractPriceInfo = await this.minter
+          .connect(this.accounts.owner)
+          .getPriceInfo(projectOne);
+        await ethers.provider.send("evm_mine", [
+          this.startTime + auctionStartTimeOffset + i * step,
         ]);
         await this.minter.connect(this.accounts.owner).purchase(projectOne, {
           value: price.toString(),
@@ -185,7 +193,8 @@ describe("MinterDAExpV1_V1Core", async function () {
         let ownerDelta = (await this.accounts.owner.getBalance()).sub(
           ownerBalance
         );
-        expect(ownerDelta.mul("-1").lte(price)).to.be.true;
+        expect(ownerDelta.mul("-1").lte(contractPriceInfo.tokenPriceInWei)).to
+          .be.true;
       }
     });
 
@@ -197,8 +206,8 @@ describe("MinterDAExpV1_V1Core", async function () {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           projectOne,
-          this.startTime + auctionStartTimeOffset,
-          defaultHalfLife,
+          this.startTime + ONE_HOUR,
+          this.startTime + 2 * ONE_HOUR,
           startingPrice,
           basePrice
         );
@@ -217,13 +226,15 @@ describe("MinterDAExpV1_V1Core", async function () {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           projectOne,
-          this.startTime + auctionStartTimeOffset,
-          defaultHalfLife,
+          this.startTime + ONE_HOUR,
+          this.startTime + 2 * ONE_HOUR,
           startingPrice,
           basePrice
         );
 
-      await ethers.provider.send("evm_mine", [this.startTime + 5 * ONE_HOUR]);
+      await ethers.provider.send("evm_mine", [
+        this.startTime + auctionStartTimeOffset + 2 * ONE_HOUR,
+      ]);
 
       let contractPriceInfo = await this.minter
         .connect(this.accounts.owner)
@@ -246,13 +257,14 @@ describe("MinterDAExpV1_V1Core", async function () {
 
       const receipt = await ethers.provider.getTransactionReceipt(tx.hash);
       const txCost = receipt.effectiveGasPrice.mul(receipt.gasUsed).toString();
+
       console.log(
-        "Gas cost for a successful Exponential DA mint: ",
+        "Gas cost for a successful Linear DA mint: ",
         ethers.utils.formatUnits(txCost, "ether").toString(),
         "ETH"
       );
 
-      expect(txCost.toString()).to.equal(ethers.utils.parseEther("0.0373542")); // assuming a cost of 100 GWEI
+      expect(txCost.toString()).to.equal(ethers.utils.parseEther("0.0373481")); // assuming a cost of 100 GWEI
     });
   });
 
@@ -298,16 +310,14 @@ describe("MinterDAExpV1_V1Core", async function () {
 
   describe("setAuctionDetails", async function () {
     it("cannot be modified mid-auction", async function () {
-      await ethers.provider.send("evm_mine", [
-        this.startTime + 2 * auctionStartTimeOffset,
-      ]);
+      await ethers.provider.send("evm_mine", [this.startTime + ONE_HOUR]);
       await expectRevert(
         this.minter
           .connect(this.accounts.artist)
           .setAuctionDetails(
             projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
+            this.startTime + ONE_MINUTE,
+            this.startTime + 2 * ONE_HOUR,
             startingPrice,
             basePrice
           ),
@@ -323,8 +333,8 @@ describe("MinterDAExpV1_V1Core", async function () {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           projectOne,
-          this.startTime + auctionStartTimeOffset,
-          defaultHalfLife,
+          this.startTime + ONE_MINUTE,
+          this.startTime + 2 * ONE_HOUR,
           startingPrice,
           basePrice
         );
@@ -339,8 +349,8 @@ describe("MinterDAExpV1_V1Core", async function () {
           .connect(this.accounts.additional)
           .setAuctionDetails(
             projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
+            this.startTime + ONE_MINUTE,
+            this.startTime + 2 * ONE_HOUR,
             startingPrice,
             basePrice
           ),
@@ -355,8 +365,8 @@ describe("MinterDAExpV1_V1Core", async function () {
           .connect(this.accounts.deployer)
           .setAuctionDetails(
             projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
+            this.startTime + ONE_MINUTE,
+            this.startTime + 2 * ONE_HOUR,
             startingPrice,
             basePrice
           ),
@@ -373,8 +383,8 @@ describe("MinterDAExpV1_V1Core", async function () {
           .connect(this.accounts.artist)
           .setAuctionDetails(
             projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
+            this.startTime + ONE_MINUTE,
+            this.startTime + 2 * ONE_HOUR,
             basePrice,
             startingPrice
           ),
@@ -440,78 +450,63 @@ describe("MinterDAExpV1_V1Core", async function () {
     });
   });
 
-  describe("enforce and broadcasts auction half-life", async function () {
-    it("enforces half-life min/max constraint", async function () {
+  describe("enforce and broadcasts min auction length", async function () {
+    it("enforces min/max auction length constraint", async function () {
+      await this.minter
+        .connect(this.accounts.deployer)
+        .resetAuctionDetails(projectOne);
+      // expect revert when creating a new project with min/max reversed
+      await expectRevert(
+        this.minter
+          .connect(this.accounts.artist)
+          .setAuctionDetails(
+            projectOne,
+            this.startTime + ONE_HOUR * 2,
+            this.startTime + ONE_HOUR,
+            startingPrice,
+            basePrice
+          ),
+        "Auction end must be greater than auction start"
+      );
+    });
+
+    it("enforces min auction length constraint", async function () {
       await this.minter
         .connect(this.accounts.deployer)
         .resetAuctionDetails(projectOne);
       // expect revert when creating a new project with
-      const invalidHalfLifeSecondsMin = ONE_MINUTE;
+      const invalidLengthSeconds = 60;
       await expectRevert(
         this.minter
           .connect(this.accounts.artist)
           .setAuctionDetails(
             projectOne,
-            this.startTime + auctionStartTimeOffset,
-            invalidHalfLifeSecondsMin,
+            this.startTime + ONE_HOUR,
+            this.startTime + ONE_HOUR + invalidLengthSeconds,
             startingPrice,
             basePrice
           ),
-        "Price decay half life must fall between min and max allowable values"
-      );
-
-      // expect revert when creating a new project with
-      const invalidHalfLifeSecondsMax = ONE_DAY;
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.artist)
-          .setAuctionDetails(
-            projectOne,
-            this.startTime + auctionStartTimeOffset,
-            invalidHalfLifeSecondsMax,
-            startingPrice,
-            basePrice
-          ),
-        "Price decay half life must fall between min and max allowable values"
+        "Auction length must be at least minimumAuctionLengthSeconds"
       );
     });
 
-    it("emits event when allowable half life range is updated", async function () {
-      const newMinSeconds = 60;
-      const newMaxSeconds = 6000;
-      // emits event when allowable half life range is updated
+    it("emits event when min auction length is updated", async function () {
+      const newLengthSeconds = 3601;
+      // emits event when minimum auction length is updated
       await expect(
         this.minter
           .connect(this.accounts.deployer)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(
-            newMinSeconds,
-            newMaxSeconds
-          )
+          .setMinimumAuctionLengthSeconds(newLengthSeconds)
       )
-        .to.emit(this.minter, "AuctionHalfLifeRangeSecondsUpdated")
-        .withArgs(newMinSeconds, newMaxSeconds);
+        .to.emit(this.minter, "MinimumAuctionLengthSecondsUpdated")
+        .withArgs(newLengthSeconds);
     });
 
-    it("validate setAllowablePriceDecayHalfLifeRangeSeconds guards", async function () {
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.deployer)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(600, 60),
-        "Maximum half life must be greater than minimum"
-      );
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.deployer)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(0, 600),
-        "Half life of zero not allowed"
-      );
-    });
-
-    it("validate setAllowablePriceDecayHalfLifeRangeSeconds ACL", async function () {
+    it("validate setMinimumAuctionLengthSeconds ACL", async function () {
       await expectRevert(
         this.minter
           .connect(this.accounts.additional)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(60, 600),
+          .setMinimumAuctionLengthSeconds(600),
         "Only Core whitelisted"
       );
     });
@@ -557,6 +552,7 @@ describe("MinterDAExpV1_V1Core", async function () {
 
   describe("reentrancy attack", async function () {
     it("does not allow reentrant purchaseTo", async function () {
+      // advance to time when auction is active
       await ethers.provider.send("evm_mine", [
         this.startTime + auctionStartTimeOffset,
       ]);
