+++ test/MinterSuiteMinters/MinterSetPrice/MinterSetPriceV1.PRTNR.test.ts
@@ -12,47 +12,49 @@ import { ethers } from "hardhat";
 import EthersAdapter from "@gnosis.pm/safe-ethers-lib";
 import Safe from "@gnosis.pm/safe-core-sdk";
 import { SafeTransactionDataPartial } from "@gnosis.pm/safe-core-sdk-types";
-import { getGnosisSafe } from "./util/GnosisSafeNetwork";
+import { getGnosisSafe } from "../../util/GnosisSafeNetwork";
 
 /**
  * These tests intended to ensure this Filtered Minter integrates properly with
- * V1 core contract.
+ * V2 PRTNR core contract.
  */
-describe("MinterSetPriceV1_V1Core", async function () {
+describe("MinterSetPriceV1_V2PRTNRCore", async function () {
   const name = "Non Fungible Token";
   const symbol = "NFT";
 
-  const firstTokenId = new BN("30000000");
-  const secondTokenId = new BN("3000001");
+  const firstTokenId = new BN("0");
+  const secondTokenId = new BN("1");
 
   const pricePerTokenInWei = ethers.utils.parseEther("1");
   const higherPricePerTokenInWei = pricePerTokenInWei.add(
     ethers.utils.parseEther("0.1")
   );
-  const projectOne = 3; // V1 core starts at project 3
-  const projectTwo = 4;
-  const projectThree = 5;
+  const projectOne = 0;
+  const projectTwo = 1;
+  const projectThree = 2;
 
   const projectMaxInvocations = 15;
 
   beforeEach(async function () {
-    const [owner, newOwner, artist, additional, snowfro] =
+    const [owner, newOwner, artist, additional, deployer] =
       await ethers.getSigners();
     this.accounts = {
       owner: owner,
       newOwner: newOwner,
       artist: artist,
       additional: additional,
-      snowfro: snowfro,
+      deployer: deployer,
     };
     const randomizerFactory = await ethers.getContractFactory(
       "BasicRandomizer"
     );
     this.randomizer = await randomizerFactory.deploy();
 
-    const artblocksFactory = await ethers.getContractFactory("GenArt721CoreV1");
-    this.token = await artblocksFactory
-      .connect(snowfro)
+    const coreFactory = await ethers.getContractFactory(
+      "GenArt721CoreV2_PRTNR"
+    );
+    this.token = await coreFactory
+      .connect(deployer)
       .deploy(name, symbol, this.randomizer.address);
 
     const minterFilterFactory = await ethers.getContractFactory(
@@ -75,23 +77,23 @@ describe("MinterSetPriceV1_V1Core", async function () {
     );
 
     await this.token
-      .connect(snowfro)
-      .addProject("project1", artist.address, 0, false);
+      .connect(deployer)
+      .addProject("project1", artist.address, 0);
 
     await this.token
-      .connect(snowfro)
-      .addProject("project2", artist.address, 0, false);
+      .connect(deployer)
+      .addProject("project2", artist.address, 0);
 
     await this.token
-      .connect(snowfro)
-      .addProject("project3", artist.address, 0, false);
+      .connect(deployer)
+      .addProject("project3", artist.address, 0);
 
-    await this.token.connect(snowfro).toggleProjectIsActive(projectOne);
-    await this.token.connect(snowfro).toggleProjectIsActive(projectTwo);
-    await this.token.connect(snowfro).toggleProjectIsActive(projectThree);
+    await this.token.connect(deployer).toggleProjectIsActive(projectOne);
+    await this.token.connect(deployer).toggleProjectIsActive(projectTwo);
+    await this.token.connect(deployer).toggleProjectIsActive(projectThree);
 
     await this.token
-      .connect(snowfro)
+      .connect(deployer)
       .addMintWhitelisted(this.minterFilter.address);
 
     await this.token
@@ -109,20 +111,20 @@ describe("MinterSetPriceV1_V1Core", async function () {
     await this.token.connect(artist).toggleProjectIsPaused(projectThree);
 
     await this.minterFilter
-      .connect(this.accounts.snowfro)
+      .connect(this.accounts.deployer)
       .addApprovedMinter(this.minter1.address);
     await this.minterFilter
-      .connect(this.accounts.snowfro)
+      .connect(this.accounts.deployer)
       .addApprovedMinter(this.minter2.address);
     await this.minterFilter
-      .connect(this.accounts.snowfro)
+      .connect(this.accounts.deployer)
       .addApprovedMinter(this.minter3.address);
 
     await this.minterFilter
-      .connect(this.accounts.snowfro)
+      .connect(this.accounts.deployer)
       .setMinterForProject(projectOne, this.minter1.address);
     await this.minterFilter
-      .connect(this.accounts.snowfro)
+      .connect(this.accounts.deployer)
       .setMinterForProject(projectTwo, this.minter2.address);
     // We leave project three with no minter on purpose
 
@@ -141,7 +143,7 @@ describe("MinterSetPriceV1_V1Core", async function () {
         "GenArt721CoreV1"
       );
       const token2 = await artblocksFactory
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .deploy(name, symbol, this.randomizer.address);
 
       const minterFilterFactory = await ethers.getContractFactory(
@@ -170,10 +172,10 @@ describe("MinterSetPriceV1_V1Core", async function () {
           .updatePricePerTokenInWei(projectOne, higherPricePerTokenInWei),
         onlyArtistErrorMessage
       );
-      // doesn't allow snowfro
+      // doesn't allow deployer
       await expectRevert(
         this.minter1
-          .connect(this.accounts.snowfro)
+          .connect(this.accounts.deployer)
           .updatePricePerTokenInWei(projectOne, higherPricePerTokenInWei),
         onlyArtistErrorMessage
       );
@@ -213,7 +215,7 @@ describe("MinterSetPriceV1_V1Core", async function () {
       const needMoreValueErrorMessage = "Must send minimum value to mint!";
       // update project two to use minter one
       await this.minterFilter
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .setMinterForProject(projectTwo, this.minter1.address);
       // artist increases price of project one
       await this.minter1
@@ -247,7 +249,7 @@ describe("MinterSetPriceV1_V1Core", async function () {
   describe("purchase", async function () {
     it("does not allow purchase prior to configuring price", async function () {
       await this.minterFilter
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .setMinterForProject(projectThree, this.minter3.address);
       await expectRevert(
         this.minter3.connect(this.accounts.owner).purchase(projectThree, {
@@ -365,14 +367,35 @@ describe("MinterSetPriceV1_V1Core", async function () {
         "ETH"
       );
 
-      expect(txCost.toString()).to.equal(ethers.utils.parseEther("0.0361817")); // assuming a cost of 100 GWEI
+      expect(txCost.toString()).to.equal(ethers.utils.parseEther("0.0216789")); // assuming a cost of 100 GWEI
+    });
+
+    it("mints and calculates gas values for project two", async function () {
+      await this.minter2
+        .connect(this.accounts.artist)
+        .updatePricePerTokenInWei(projectTwo, pricePerTokenInWei);
+      const tx = await this.minter2
+        .connect(this.accounts.owner)
+        .purchase(projectTwo, {
+          value: pricePerTokenInWei,
+        });
+
+      const receipt = await ethers.provider.getTransactionReceipt(tx.hash);
+      const txCost = receipt.effectiveGasPrice.mul(receipt.gasUsed).toString();
+      console.log(
+        "Gas cost for a successful Ether mint: ",
+        ethers.utils.formatUnits(txCost, "ether").toString(),
+        "ETH"
+      );
+
+      expect(txCost.toString()).to.equal(ethers.utils.parseEther("0.0296401")); // assuming a cost of 100 GWEI
     });
   });
 
   describe("purchaseTo", async function () {
     it("does not allow purchase prior to configuring price", async function () {
       await this.minterFilter
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .setMinterForProject(projectThree, this.minter3.address);
       await expectRevert(
         this.minter3
@@ -418,41 +441,41 @@ describe("MinterSetPriceV1_V1Core", async function () {
   });
 
   describe("setProjectMaxInvocations", async function () {
-    it("handles getting tokenInfo invocation info with V1 core", async function () {
+    it("handles getting tokenInfo invocation info with V2 core", async function () {
       await this.minter1
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .setProjectMaxInvocations(projectOne);
       // minter should update storage with accurate projectMaxInvocations
       await this.minter1
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .setProjectMaxInvocations(projectOne);
       let maxInvocations = await this.minter1
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .projectMaxInvocations(projectOne);
       expect(maxInvocations).to.be.equal(projectMaxInvocations);
       // ensure hasMaxBeenReached did not unexpectedly get set as true
       let hasMaxBeenInvoked = await this.minter1
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .projectMaxHasBeenInvoked(projectOne);
       expect(hasMaxBeenInvoked).to.be.false;
       // ensure minter2 gives same results
       await this.minter2
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .setProjectMaxInvocations(projectTwo);
       await this.minter2
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .setProjectMaxInvocations(projectTwo);
       maxInvocations = await this.minter2
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .projectMaxInvocations(projectTwo);
       expect(maxInvocations).to.be.equal(projectMaxInvocations);
       // should also support unconfigured project projectMaxInvocations
       // e.g. projectOne on minter3 - still update to accurate max invocations
       await this.minter3
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .setProjectMaxInvocations(projectOne);
       maxInvocations = await this.minter3
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .projectMaxInvocations(projectOne);
       expect(maxInvocations).to.be.equal(projectMaxInvocations);
     });
@@ -507,7 +530,7 @@ describe("MinterSetPriceV1_V1Core", async function () {
         "ReentrancyMock"
       );
       const reentrancyMock = await reentrancyMockFactory
-        .connect(this.accounts.snowfro)
+        .connect(this.accounts.deployer)
         .deploy();
       // attacker should see revert when performing reentrancy attack
       const totalTokensToMint = 2;
@@ -515,7 +538,7 @@ describe("MinterSetPriceV1_V1Core", async function () {
       let totalValue = higherPricePerTokenInWei.mul(numTokensToMint);
       await expectRevert(
         reentrancyMock
-          .connect(this.accounts.snowfro)
+          .connect(this.accounts.deployer)
           .attack(
             numTokensToMint,
             this.minter1.address,
@@ -533,7 +556,7 @@ describe("MinterSetPriceV1_V1Core", async function () {
       totalValue = higherPricePerTokenInWei.mul(numTokensToMint);
       for (let i = 0; i < totalTokensToMint; i++) {
         await reentrancyMock
-          .connect(this.accounts.snowfro)
+          .connect(this.accounts.deployer)
           .attack(
             numTokensToMint,
             this.minter1.address,
